Backend passing up to US 4
backend tests all passing up to us 4. This includes table validations for get, put and post requests. Had to change seed names as well to align with tests. Most table validation occurs inside the controller, as opposed to the middleware files I used prior, as it was causing the async timeouts.