US 6 mostly done
Back end handling of US 6 mostly complete except for the Finished status portion on the tables controller. Issue is that the tests are somehow dropping the reservation ID between steps and that's causing my program to think the table is unoccupied and therefore doesn't need to be marked as finished. Will need to test further or rope in a TA