All tests passing, frontend cleaned
All tests are now passing, which was mostly fixing up the UpdateReservation to correctly take in the date and time from the information. I also removed all console logs on the front so that I could see what was happening in the tests. Front and back are set, now it's just about adding some polish and making the pages look better.