fixed issues for revision request
Added the "finished" and "cancelled" exclusion to reservation list to only activate on the back end so that search could still find those reservations. Added a regex check to the reservation form to make sure it's a number in the proper format. Also added the formatAsDate function to the query builder on NewReservation as it was returning the full date string.